Fix bug #4 in tax_glom function
Function was producing different results from phyloseq's tax_glom. Issue
was that the call to top_n was keeping more than one OTU in cases of
ties. Fixed by picking the first OTU in cases of ties, which gives the
same taxon that would be chosen by phyloseq's call to which.max in
`merge_taxa`.